Split files in login and registration
